Added an extra column and reordered streams
The extra column shows content generated by users that you follow in
their home space (this is the old behaviour of Jon's activity stuff.

The reordering is an attempt to put the stuff that's most interesting to
you on the left and the stuff that's least interesting on the right. The
@ stream isn't on the far left though as it doesn't update often enough.